requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.4.0
distro==1.9.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.4
jiter==0.7.0
MouseInfo==0.1.3
msedge-selenium-tools==3.141.4
mss==9.0.1
multidict==6.1.0
numpy==1.24.4
openai==0.28.0
openpyxl==3.1.5
outcome==1.2.0
packaging==24.1
pandas==2.0.3
Pillow==9.5.0
propcache==0.2.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
requests==2.32.3
requests-futures==1.0.1
scapy==2.6.1
selenium==3.141.0
sherlock-project==0.15.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stem==1.8.2
tqdm==4.67.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.4
webdriver-manager==4.0.2
wsproto==1.2.0
yarl==1.15.2
